+++ das/.gitignore
@@ -1,2 +1,3 @@
 .recommenders/
 .metadata
+/bin/
